requirements.txt
Django==4.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==23.3
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
dnspython==2.4.2
python-decouple==3.8